Removed Google OAuth secrets

+++ src/main/java/com/hotelbooking/airbnb/service/impl/HotelServiceImpl.java
@@ -3,10 +3,14 @@ package com.hotelbooking.airbnb.service.impl;
 import com.hotelbooking.airbnb.dto.*;
 import com.hotelbooking.airbnb.entity.Hotel;
 import com.hotelbooking.airbnb.entity.Room;
+import com.hotelbooking.airbnb.entity.User;
+import com.hotelbooking.airbnb.exception.APIException;
 import com.hotelbooking.airbnb.exception.ResourceNotFoundException;
 import com.hotelbooking.airbnb.repository.HotelRepository;
 import com.hotelbooking.airbnb.repository.InventoryRepository;
 import com.hotelbooking.airbnb.repository.RoomRepository;
+import com.hotelbooking.airbnb.repository.UserRepository;
+import com.hotelbooking.airbnb.security.util.AuthUtil;
 import com.hotelbooking.airbnb.service.HotelService;
 import com.hotelbooking.airbnb.service.InventoryService;
 import jakarta.transaction.Transactional;
@@ -17,10 +21,12 @@ import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.time.temporal.ChronoUnit;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @Slf4j
@@ -28,23 +34,36 @@ import java.util.List;
 public class HotelServiceImpl implements HotelService {
     private final RoomRepository roomRepository;
     private final InventoryRepository inventoryRepository;
+    private final UserRepository userRepository;
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
     private final InventoryService inventoryService;
+    private final AuthUtil authUtil;
 
     @Override
     public HotelDto createNewHotel(HotelDto hotelDto) {
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
+        boolean isManager = authUtil.getPrincipal().getAuthorities().stream()
+                .anyMatch(a -> Objects.equals(a.getAuthority(), "ROLE_HOTEL_MANAGER"));
+
+        if (!isManager) {
+            throw new APIException(HttpStatus.FORBIDDEN, "Only Hotel Managers can create hotels");
+        }
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User currentUser = userRepository.getReferenceById(authUtil.loggedInUserId());
+        hotel.setOwner(currentUser);
+
         hotel = hotelRepository.save(hotel);
-        log.info("Created a new hotel with ID: {}", hotelDto.getId());
+        log.info("Created a new hotel with ID: {}", hotel.getId());
         return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
     public List<HotelDto> getAllHotels(){
         List<Hotel> hotels = hotelRepository.findAll();
+
         return hotels.stream()
                 .map(hotel -> modelMapper.map(hotel, HotelDto.class))
                 .toList();
@@ -68,6 +87,9 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel", "hotelId", id));
+
+        validateHotelOwnership(hotel);
+
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -78,6 +100,9 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel", "hotelId", id));
+
+        validateHotelOwnership(hotel);
+
         hotel.setCity(hotelDto.getCity());
         hotel.setName(hotelDto.getName());
         hotel.setAmenities(hotelDto.getAmenities());
@@ -93,6 +118,8 @@ public class HotelServiceImpl implements HotelService {
     public void deleteHotelById(Long hotelId) {
         Hotel hotel = hotelRepository.findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel", "hotelId", hotelId));
+
+        validateHotelOwnership(hotel);
 
         for (Room room : hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
@@ -112,11 +139,17 @@ public class HotelServiceImpl implements HotelService {
 
         log.info("Activating the hotel: {}", hotel.getActive());
 
-
         for(Room room : hotel.getRooms()){
             if (!inventoryRepository.existsByRoom(room)) {
                 inventoryService.initializeRoomForYear(room);
             }
+        }
+    }
+
+    private void validateHotelOwnership(Hotel hotel) {
+        Long loggedInUserId = authUtil.loggedInUserId();
+        if (!hotel.getOwner().getId().equals(loggedInUserId)) {
+            throw new APIException(HttpStatus.FORBIDDEN, "Access denied: You are not the owner of this hotel.");
         }
     }
 }
